+++ packages/agent/src/utils/forest-schema/generator-action-field-widget.ts
@@ -11,6 +11,7 @@ import {
   ActionFieldRichText,
   ActionFieldTextArea,
   ActionFieldTextInputList,
+  ActionFieldUserDropdown,
 } from '@forestadmin/datasource-toolkit';
 import {
   ForestServerActionField,
@@ -29,6 +30,8 @@ import {
   ForestServerActionFieldTextAreaOptions,
   ForestServerActionFieldTextInputListOptions,
   ForestServerActionFieldTextInputOptions,
+  ForestServerActionFieldTimePickerOptions,
+  ForestServerActionFieldUserDropdown,
 } from '@forestadmin/forestadmin-client';
 
 import ActionFields from './action-fields';
@@ -76,6 +79,12 @@ export default class GeneratorActionFieldWidget {
 
     if (ActionFields.isCurrencyInputField(field))
       return GeneratorActionFieldWidget.buildCurrencyInputWidgetEdit(field);
+
+    if (ActionFields.isUserDropdownField(field))
+      return GeneratorActionFieldWidget.buildUserDropdownWidgetEdit(field);
+
+    if (ActionFields.isTimePicker(field))
+      return GeneratorActionFieldWidget.buildTimePickerWidgetEdit();
 
     if (ActionFields.isJsonEditorField(field))
       return GeneratorActionFieldWidget.buildJsonEditorWidgetEdit();
@@ -249,6 +258,13 @@ export default class GeneratorActionFieldWidget {
     };
   }
 
+  private static buildTimePickerWidgetEdit(): ForestServerActionFieldTimePickerOptions {
+    return {
+      name: 'time editor',
+      parameters: {},
+    };
+  }
+
   private static buildCurrencyInputWidgetEdit(
     field: ActionFieldCurrencyInput,
   ): ForestServerActionFieldCurrencyInputOptions {
@@ -279,6 +295,17 @@ export default class GeneratorActionFieldWidget {
       default:
         return 'Unit';
     }
+  }
+
+  private static buildUserDropdownWidgetEdit(
+    field: ActionFieldUserDropdown,
+  ): ForestServerActionFieldUserDropdown {
+    return {
+      name: 'assignee editor',
+      parameters: {
+        placeholder: field.placeholder || null,
+      },
+    };
   }
 
   private static isValidNumber(value: unknown): boolean {