+++ packages/datasource-toolkit/src/interfaces/action.ts
@@ -34,6 +34,7 @@ export const ActionFieldTypeList = [
   'Collection',
   'Date',
   'Dateonly',
+  'Time',
   'Enum',
   'File',
   'Json',
@@ -162,6 +163,17 @@ export type ActionFieldCurrencyInput = ActionFieldBase & {
   base?: 'Unit' | 'Cent';
 };
 
+export type ActionFieldUserDropdown = ActionFieldBase & {
+  type: 'String';
+  widget: 'UserDropdown';
+  placeholder?: string;
+};
+
+export type ActionFieldTimePicker = ActionFieldBase & {
+  type: 'Time';
+  widget: 'TimePicker';
+};
+
 export type ActionFieldJsonEditor = ActionFieldBase & {
   type: 'Json';
   widget: 'JsonEditor';
@@ -203,6 +215,8 @@ export type ActionField = StrictUnion<
   | ActionFieldNumberInputList
   | ActionFieldDatePickerInput
   | ActionFieldCurrencyInput
+  | ActionFieldUserDropdown
+  | ActionFieldTimePicker
   | ActionFieldJsonEditor
   | ActionFieldAddressPicker
 >;
@@ -215,13 +229,16 @@ export type ActionFieldWidget =
   | 'TextInput'
   | 'TextInputList'
   | 'TextArea'
+  | 'Timepicker'
   | 'RichText'
   | 'NumberInput'
   | 'NumberInputList'
   | 'ColorPicker'
   | 'DatePicker'
-  | 'JsonEditor'
-  | 'AddressPicker';
+  | 'AddressPicker'
+  | 'UserDropdown'
+  | 'TimePicker'
+  | 'JsonEditor';
 
 export type SuccessResult = {
   type: 'Success';