+++ packages/plugin-flattener/CHANGELOG.md
@@ -1,3 +1,25 @@
+## @forestadmin/plugin-flattener [1.0.57](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.56...@forestadmin/plugin-flattener@1.0.57) (2023-09-25)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.26.0
+* **@forestadmin/datasource-customizer:** upgraded to 1.33.0
+
+## @forestadmin/plugin-flattener [1.0.56](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.55...@forestadmin/plugin-flattener@1.0.56) (2023-09-25)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.25.0
+* **@forestadmin/datasource-customizer:** upgraded to 1.32.0
+
 ## @forestadmin/plugin-flattener [1.0.55](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.54...@forestadmin/plugin-flattener@1.0.55) (2023-09-22)